ml_model_deploy/requirements.txt
flask==2.2.2
pandas==1.5.3
requests==2.28.2
scikit-learn
psycopg2-binary
lottie